Create CalculateSpeedStatisticsSpecies
Calculates speed statistics (mean, min, max, median, st dev) for each species of animal given a gdb of that species satellite tracks. The outputs are the overall statistics for the species and also a text file containing the stats for each individual animal. The text file can be imported to Excel to visualize the average speeds as a histogram to identify a speed threshold for speed filtering.